requirements.txt
pytest
requests
selenium
allure-pytest
